Adjuster Folder Structure
