--- Example/Example Invoker/Example Invoker/App.config
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/>
-    </startup>
-</configuration>
