--- Example/Example Invoker/Example Invoker/Example Invoker.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Example Invoker", "Example Invoker.csproj", "{A08A747C-D964-44CC-8A53-E9EF2EA74241}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A08A747C-D964-44CC-8A53-E9EF2EA74241}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A08A747C-D964-44CC-8A53-E9EF2EA74241}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A08A747C-D964-44CC-8A53-E9EF2EA74241}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A08A747C-D964-44CC-8A53-E9EF2EA74241}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
