+++ examples/Example Invoker/Example Invoker/Form1.resx
