+++ examples/Example Invoker/Example Invoker/Properties/AssemblyInfo.cs
