+++ examples/Example Invoker/Example Invoker/Properties/Resources.resx
