+++ examples/Example Invoker/Example Invoker/Properties/Settings.settings
