+++ examples/Example Invoker/Example Invoker/bin/Debug/Example Invoker.vshost.exe.config
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<configuration>
+    <startup> 
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/>
+    </startup>
+</configuration>
