+++ src/MyLDC/MyLDC.depend
@@ -1,0 +1,29 @@
+# depslib dependency file v1.0
+1477607307 source:c:\users\utente\documents\cpp\myldc\myldc\myldc\main.cpp
+	"main.h"
+
+1477607608 c:\users\utente\documents\cpp\myldc\myldc\myldc\main.h
+	<windows.h>
+	<stdio.h>
+	<sstream>
+
+1477645109 source:c:\programmi\git\myldc\myldc\main.cpp
+	"main.h"
+
+1477645537 c:\programmi\git\myldc\myldc\main.h
+	<windows.h>
+	<stdio.h>
+	<sstream>
+	<iostream>
+	<fstream>
+
+1477640116 source:c:\programmi\git\myldc\myldc\database.cpp
+	"database.h"
+
+1477640020 c:\programmi\git\myldc\myldc\database.h
+	<iostream>
+	<fstream>
+	<windows.h>
+	<stdio.h>
+	<sstream>
+
