+++ src/MyLDC/bin/Debug/libMyLDC.def
@@ -1,0 +1,10 @@
+EXPORTS
+    _ZN4RowsC1Ev @1
+    _ZN4RowsD1Ev @2
+    _ZN5MyLDC11CreateTableESsSs @3
+    _ZN5MyLDC11InfoSistemaEv @4
+    _ZN5MyLDC8CreateDbEv @5
+    _ZStorSt13_Ios_OpenmodeS_ @6
+    _get_output_format @7
+    row @8 DATA
+    rw @9 DATA
